Split test file into corner cases and type-specific ones
